Task: Edit the HotCollections to the specified requirements
Why: Approaching the project in order of
what is necessary to have a funcitonal NFT
Marketplace.

How: Adjusted the App.jsx, Hotcollections.jsx,
style.css and made major changes in order to
include an API request for appropriate displayed
information in the hotcollections section
as well as added a loading state and responsivenes
for the different screen sizes, and added the
NFT id display in the itemDetails url screen.